% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_compound_interest.R
\name{calc_compound_interest}
\alias{calc_compound_interest}
\title{Calculate Compound Interest}
\usage{
calc_compound_interest(P, r, n, t)
}
\arguments{
\item{P}{Principal amount}

\item{r}{Annual interest rate (in decimal form, e.g., 5% = 0.05)}

\item{n}{Number of times the interest is compounded per year}

\item{t}{Time in years}
}
\value{
The amount of money accumulated after interest
}
\description{
This function calculates compound interest based on principal, interest rate,
the number of times interest is compounded per year, and the time in years.
}
\examples{
calc_compound_interest(1000, 0.05, 4, 5)
}
