% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pay_off_debt.R
\name{calc_debt_payoff}
\alias{calc_debt_payoff}
\title{Calculate Time to Pay Off Debt}
\usage{
calc_debt_payoff(P, M, r)
}
\arguments{
\item{P}{Principal amount (loan amount)}

\item{M}{Monthly payment}

\item{r}{Monthly interest rate (annual rate / 12)}
}
\value{
The number of months to pay off the debt
}
\description{
This function calculates how many months it would take to pay off a debt given
the principal amount, monthly payment, and monthly interest rate.
}
\examples{
calc_debt_payoff(5000, 200, 0.01)
}
