% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_payment.R
\name{calc_loan_payment}
\alias{calc_loan_payment}
\title{Calculate Loan Payment}
\usage{
calc_loan_payment(P, r, n)
}
\arguments{
\item{P}{Principal amount (loan amount)}

\item{r}{Monthly interest rate (annual rate / 12)}

\item{n}{Number of months (loan term in years * 12)}
}
\value{
The monthly payment
}
\description{
This function calculates the monthly payment required to pay off a loan.
}
\examples{
calc_loan_payment(5000, 0.01, 36)
}
