% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_X_amount.R
\name{calc_savings_time}
\alias{calc_savings_time}
\title{Calculate Time to Save X Amount}
\usage{
calc_savings_time(P, M, r, target)
}
\arguments{
\item{P}{Initial savings amount}

\item{M}{Monthly deposit}

\item{r}{Monthly interest rate (annual rate / 12)}

\item{target}{Target amount to save}
}
\value{
The number of months to save the target amount
}
\description{
This function calculates how many months it would take to save a certain amount of money
given an initial savings amount, monthly deposit, and monthly interest rate.
}
\examples{
calc_savings_time(1000, 200, 0.01, 10000)
}
